[enhance] For serialization run the reverse check to make sure object produces same value
This only really helps when transient is marked on object in class and a
constructor happened to have created the object.  In that case, the
transient will be null rather than object type.  Serialization occurred
just not fully as expected.  This is an edge case situation at least at
time of testing.  More likely this would happen at runtime when objects
are added and manipulated more than can be done here.