[delete] Remove test element that does nothing
this is not injected as the framework 'newing' the object.  Even if I move this to the only constructor it still does nothing so the indication that it breaks equals is wrong.  that likely existed before it was patched to skip deprecated constructor testing which was the purpose of adding dates to start with.