[enhance] Fix long standing bug with date object
Since we are not actually testing date but rather the object, it has a getMonth / setMonth that takes our provided value which is 1 in this case and processes that.  Under the hood, date object getter/setter is deprecated and also performs normalization which affects end of month / start of month due to time zone differences making it generally hard to catch.