[enhance] Boolean wrapper in java metro incorrectly uses is method instead of getter
This is a bug in java metro stack but could be anywhere.  If the standard getter is missing but setter exists, try to use 'is' to find matching method.  This may cause extra processing as this current patch does not confirm its a boolean type but rather just assumes that might be condition as to why getter is missing and uses reflection to find it.  Further, bean copier methods needed to have boolean converter used as a result.